refactor: replace setVdTotalError to helper file
